quick sort Hoare's method
The QuickSort algorithm is a significantly improved version of the 
of the exchange sort algorithm (bubbles), which is known for its low efficiency. 
The fundamental difference is that first of all, the following is swapped 
elements that are located at the maximum possible distance; after each pass 
the elements are divided into two independent groups. It is interesting to note that the improvement of the least 
the least efficient direct sorting method resulted in one of the most efficient 
of the improved methods.